[Build] Fix cross-platform build issues
- Add DXC include path for xenia-ui-d3d12 (WinAdapter.h resolution)
- Fix size_t to uint32_t conversion warning in memory.cc
- Define IR_RUNTIME_METALCPP for metal-cpp compatibility

+++ src/xenia/memory.cc
@@ -2275,8 +2275,9 @@ bool PhysicalHeap::TriggerCallbacks(
               unprotect_page = false;
             }
           } else {
-            guest_page_last_exclusive = std::min<uint32_t>(
-                guest_page_last_exclusive, page_table_.size());
+            guest_page_last_exclusive =
+                std::min<uint32_t>(guest_page_last_exclusive,
+                                   static_cast<uint32_t>(page_table_.size()));
             bool any_read_write = false;
             for (uint32_t guest_page_number = guest_page_first;
                  guest_page_number < guest_page_last_exclusive;
